api-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
google-generativeai==0.3.2
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
pandas==2.1.4
plotly==5.17.0 
python-multipart==0.0.19
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.24.4
kafka-python==2.0.2